# docs/developer-docs/public/indexer/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tari_indexer",
    "description": "The Tari indexer implementation",
    "contact": {
      "name": "The Tari Development Community"
    },
    "license": {
      "name": "BSD-3-Clause",
      "identifier": "BSD-3-Clause"
    },
    "version": "0.37.0"
  },
  "paths": {
...
    "/templates/cached": {
      "get": {
        "tags": [
          "handlers::templates"
        ],
        "description": "List all template cached by this indexer",
        "operationId": "list_cached_templates",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Limit the number of results returned",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of cached templates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListTemplatesResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to list cached templates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/templates/catalogue": {
      "get": {
        "tags": [
          "handlers::templates"
        ],
        "description": "List templates discovered on the network via the template catalogue",
        "operationId": "list_template_catalogue",
        "parameters": [
          {
            "name": "name_filter",
            "in": "query",
            "description": "Substring filter on template name",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum entries to return (default: 20, max: 100)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "after",
            "in": "query",
            "description": "Cursor: return entries inserted after the row with this template address. Omit to start from the beginning",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Template catalogue entries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListTemplateCatalogueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to list template catalogue",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/templates/catalogue/{template_address}": {
      "get": {
        "tags": [
          "handlers::templates"
        ],
        "description": "Get a single template catalogue entry by its address",
        "operationId": "get_template_catalogue_entry",
        "responses": {
          "200": {
            "description": "Template catalogue entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateCatalogueItem"
                }
              }
            }
          },
          "404": {
            "description": "Template not found in catalogue",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch template catalogue entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/templates/{template_address}": {
      "get": {
        "tags": [
          "handlers::templates"
        ],
        "description": "Fetch a template definition by its address",
        "operationId": "get_template_definition",
        "responses": {
          "200": {
            "description": "Template definition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTemplateDefinitionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Template not found or still pending",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch template definition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "GetTemplateDefinitionResponse": {
        "type": "object",
        "required": [
          "name",
          "definition",
          "code_size"
        ],
        "properties": {
          "code_size": {
            "type": "integer",
            "minimum": 0
          },
          "definition": {
            "type": "object"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "ListTemplateCatalogueResponse": {
        "type": "object",
        "required": [
          "entries"
        ],
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateCatalogueItem"
            }
          }
        }
      },
      "ListTemplatesResponse": {
        "type": "object",
        "required": [
          "templates"
        ],
        "properties": {
          "templates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateMeta"
            }
          }
        }
      },
...
      "TemplateCatalogueItem": {
        "type": "object",
        "required": [
          "template_address",
          "template_name",
          "author_public_key",
          "binary_hash",
          "at_epoch"
        ],
        "properties": {
          "at_epoch": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "author_public_key": {
            "type": "string"
          },
          "binary_hash": {
            "type": "string"
          },
          "metadata_hash": {
            "type": [
              "string",
              "null"
            ]
          },
          "template_address": {
            "type": "string"
          },
          "template_name": {
            "type": "string"
          }
        }
      },
      "TemplateMeta": {
        "type": "object",
        "required": [
          "name",
          "address",
          "binary_sha",
          "author_public_key",
          "code_size",
          "epoch"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "author_public_key": {
            "type": "string"
          },
          "binary_sha": {
            "type": "string"
          },
          "code_size": {
            "type": "integer",
            "minimum": 0
          },
          "epoch": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "metadata_hash": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional multihash of off-chain CBOR metadata"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  }
}